// src/common/enums/account/permission.account.enum.ts
export enum EAccountPermissions {
    MANAGE_ACCOUNTS = 'manage_accounts',
    MANAGE_CLIENTS = 'manage_clients',
    MANAGE_SESSIONS = 'manage_sessions',
    MANAGE_TOKENS = 'manage_tokens',
    ACCESS_ANALYTICS = 'access_analytics',
    ACCESS_BASIC_FEATURES = 'access_basic_features',
    VIEW_REPORTS = 'view_reports',
    VIEW_OWN_DATA = 'view_own_data',
    READ_ONLY_ACCESS = 'read_only_access',
}